code cleaning commenting and renaming
