--- zakatcalculator/src/Features/zakatForm/Category.js
@@ -1,14 +1,0 @@
-import React from 'react'
-
-
-export default function Category(props) {
-   
-  return (
-      <div className='expand-btn'>
-           {props.Category}
-
-      </div>
-    
-        
-  )
-}
