+++ zakatcalculator/src/Features/zakatForm/ZakatTable.js
@@ -1,6 +1,5 @@
-import React from 'react'
-
-import FilterableProdcutTable from './FilterableProductTable'
+import React from 'react';
+import FilterableProdcutTable from './ZakatForm'
 import {ContextProvider} from './Context';
 
 export default function ZakatTable() {
